Move pkg/sql/lexer*.go to pkg/parser
