+++ pkg/parser/lexer.go
@@ -11,9 +11,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package sql
+package parser
 
 import (
+	"log"
 	"regexp"
 	"strings"
 	"unicode"
